Fixes bugs, adds support for appending LEEP-pathwasys data to prior runs

+++ include/application_modules.rb
@@ -1,4 +1,4 @@
-
+ 
 # file for holding functions related to specific applications - e.g. BC step code
 
 
@@ -117,13 +117,12 @@ module LEEPPathways
     "Opt-FoundationSlabBelowGrade:+:xps4inEffR20" => "lpSBG_02" , 
     "Opt-FoundationSlabBelowGrade:+:xps2inEffR10" => "lpSBG_03" , 
 
-
-
-
-
     "Opt-HVACSystem:+:gas-furnace-psc"     => "lpH&C_01", 
     "Opt-HVACSystem:+:gas-furnace-ecm"     => "lpH&C_02", 
     "Opt-HVACSystem:+:gas-furnace-ecm+AC"  => "lpH&C_03", 
+    "Opt-HVACSystem:+:CCASHP"  => "lpH&C_04", 
+    "Opt-HVACSystem:+:ASHP"  => "lpH&C_05", 
+    "Opt-HVACSystem:+:elec-baseboard"  => "lpH&C_06", 
 
     "Opt-GenericWall_1Layer_definitions:+:NC_R-17(eff)_2x6-16inOC_R22-batt_poly_vb" => "lpMWL_01",
     "Opt-GenericWall_1Layer_definitions:+:NC_R-16(eff)_2x6-16inOC_R19-batt_poly_vb" =>  "lpMWL_02",
@@ -135,10 +134,11 @@ module LEEPPathways
 
     "Opt-DHWSystem:+:gas_storagetank_w/powervent_ef0.67" => "lpDHW_01",
     "Opt-DHWSystem:+:GasInstantaneous" => "lpDHW_02",
-
+    "Opt-DHWSystem:+:HPHotWater" => "lpDHW_03",
+    "Opt-DHWSystem:+:elec_storage_ef0.89" => "lpDHW_04",
     "Opt-HRVonly:+:HRV_sre_60"         => "lpHRV_01",  
     "Opt-HRVonly:+:HRV_sre_78"         => "lpHRV_02",
-
+    "Opt-DWHRSystem:+:NA"     => "lpDWHR_00",
     "Opt-DWHRSystem:+:DWHR-eff-55"     => "lpDWHR_01"
   }
   GoodArchFields = [
@@ -198,7 +198,7 @@ module LEEPPathways
   GoodECMFields = Array.new [
     "ecmID",
     "attribute",
-    "measure"
+    "measure", 
   ]
 
   GoodLocFields = Array.new [ 
@@ -226,7 +226,7 @@ module LEEPPathways
     "Opt-HRVonly",
     "Opt-HVACSystem",
     "Opt-DHWSystem",
-    "Opt-DWHR-System",
+    "Opt-DWHRSystem",
 
   ]
 
@@ -278,7 +278,7 @@ module LEEPPathways
   #  "Opt-HRVonly",
   #  "Opt-HVACSystem",
   #  "Opt-DHWSystem",
-  #  "Opt-DWHR-System",
+  #  "Opt-DWHRSystem",
     #  "HVAC:ASHP:capacity_kW",
     #  "HVAC:ASHP:count",
     #  "HVAC:AirConditioner:capacity_kW",
@@ -358,15 +358,73 @@ module LEEPPathways
 
   end 
 
+  def self.ReconstituteDataStructure(txt)
+    firstRow = true 
+    header = Array.new 
+    data = Array.new
+    dataStructure = Array.new 
+    txt.split("\n").each do | line |
+      
+      if firstRow then 
+        header = line.split(",")
+        firstRow = false 
+      else 
+        data = line.split(",")
+        record = Hash.new 
+        i = 0 
+        until i == header.length do 
+          record[header[i]] = data[i]
+          i+=1       
+        end 
+        dataStructure.push record    
+      end 
+ 
+    end 
+    data.clear
+    header.clear
+    txt.clear 
+    return dataStructure
+  end 
 
-  def LEEPPathways.OpenOutputFiles()
-    log_out("Opening LEEP output files\n")
-    $ECM_outfile = File.open('./Pathways_ListOfECMs.csv', 'w')
-    $ARC_outfile = File.open('./Pathways_HTAPArchetypeData.csv', 'w')
-    $LOC_outfile = File.open('./Pathways_HTAPLocationData.csv', 'w')
-    $RUN_outfile = File.open('./Pathways_HTAPRunData.csv', 'w')
-    $LEEPprintHeader = true 
-    $LEEPRunID = 0 
+  def LEEPPathways.OpenOutputFiles(mode="overwrite")
+
+    log_out("Opening LEEP output files [mode=#{mode}\n")
+    if (mode == "overwrite")
+      $ECM_outfile = File.open('./Pathways_ListOfECMs.csv', 'w')
+      $ARC_outfile = File.open('./Pathways_HTAPArchetypeData.csv', 'w')
+      $LOC_outfile = File.open('./Pathways_HTAPLocationData.csv', 'w')
+      $RUN_outfile = File.open('./Pathways_HTAPRunData.csv', 'w')
+      $LEEPprintHeader = true 
+      $LEEPRunID = 0 
+    elsif ( mode =="append")
+      #begin 
+        # Parse ECM output file, and reconsititute data hash
+
+        log_out ("Rebuilding LEEP pathways data structures from prior runs")
+        previousData = File.read('./Pathways_ListOfECMs.csv')
+        $LEEPecmData = self.ReconstituteDataStructure(previousData)
+        previousData.clear
+        $LEEPecmData.each do | ecmRecord |
+          ecmRecord["searchTextKey"] = "#{ecmRecord["attribute"]}:+:#{ecmRecord["measure"]}"
+        end 
+        previousData = File.read('./Pathways_HTAPLocationData.csv')
+        $LEEPlocData = self.ReconstituteDataStructure(previousData)
+        previousData.clear 
+
+        previousData = File.read('./Pathways_HTAPArchetypeData.csv')
+        $LEEParchetypeData = self.ReconstituteDataStructure(previousData)
+        previousData.clear
+
+        $LEEPRunID = File.read('./Pathways_HTAPRunData.csv').split("\n").length
+
+        $ECM_outfile = File.open('./Pathways_ListOfECMs.csv', 'a')
+        $ARC_outfile = File.open('./Pathways_HTAPArchetypeData.csv', 'a')
+        $LOC_outfile = File.open('./Pathways_HTAPLocationData.csv', 'a')
+        $RUN_outfile = File.open('./Pathways_HTAPRunData.csv', 'a')
+        $LEEPprintHeader = false 
+    else
+      fatalerror ("Logic error - unknown mode #{mode}")
+    end
   end   
 
   def LEEPPathways.ExportPathwayData()
@@ -420,10 +478,10 @@ module LEEPPathways
           archID = archetype["archID"]
           foundArch = true 
           #debug_on
-          debug_out "TEST: - #{locID} [#{archetype["listOfLocations"]}]\n"
-          if (! bListContainsID(locID, archetype["listOfLocations"])) then 
-            archetype["listOfLocations"] += "#{locID};"
-          end
+          #debug_out "TEST: - #{locID} [#{archetype["listOfLocations"]}]\n"
+          #if (! bListContainsID(locID, archetype["listOfLocations"])) then 
+          #  archetype["listOfLocations"] += "#{locID};"
+          #end
           break
         end 
         debug_out ("#{foundArch}\n")
